Update to use angular cli v6 library builds and include demo app
