--- bower.json
@@ -1,44 +1,0 @@
-{
-  "name": "angular2-grid",
-  "version": "2.1.0",
-  "homepage": "https://github.com/BTMorton/angular2-grid",
-  "authors": [
-    "Ben Morton <ben.morton91@gmail.com> (http://bmorton.co.uk/)",
-    "Martin Sikora (https://github.com/martinsik)",
-    "Daniel Pastor de la Fuenta (https://github.com/aebsubis)",
-    "Frederik Schubert (https://github.com/frederikschubert)",
-    "Matthew de Nobrega (https://github.com/matthewdenobrega)",
-    "Alex (https://github.com/lifecoderua)",
-    "pocmanu (https://github.com/pocmanu)",
-    "Chris Morabito (https://github.com/morabitowoolpert)",
-    "Carlos Esteban Lopez Jaramillo (https://github.com/Luchillo)",
-    "Niklas Berg (https://github.com/nkholski)",
-    "Wouter (https://github.com/wuhkuh)",
-    "Brian Ellis (https://github.com/OnlyAGhost)"
-  ],
-  "description": "Grid-based drag/drop/resize directive plugin for Angular 2",
-  "main": "NgGrid.ts",
-  "moduleType": [
-    "umd"
-  ],
-  "keywords": [
-    "angular",
-    "angular2",
-    "angularjs",
-    "grid",
-    "drag",
-    "drop",
-    "resize",
-    "web-components"
-  ],
-  "license": "MIT",
-  "ignore": [
-    "*",
-    "!src/NgGrid.ts",
-    "!dist/NgGrid.*"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git://github.com/BTMorton/angular2-grid.git"
-  }
-}
