+++ package.json
@@ -1,66 +1,57 @@
 {
-  "name": "angular2-grid",
-  "private": true,
-  "description": "A grid-based drag/drop/resize directive plugin for Angular 2",
-  "scripts": {
-    "cleanup": "rimraf dist/bundles dist/components dist/directives dist/interfaces dist/modules dist/main.d.ts dist/main.metadata.json dist/main.js dist/main.js.map dist/LICENSE dist/README.md",
-    "bundling": "rollup -c",
-    "minify": "uglifyjs dist/bundles/NgGrid.umd.js --screw-ie8 --compress --mangle --comments --output dist/bundles/NgGrid.umd.min.js",
-    "copy": "copyfiles -f LICENSE README.md src/*.css dist",
-    "build": "npm run lint && ngc && npm run bundling && npm run minify",
-    "lint": "tslint -c tslint.json --project tsconfig.json",
-    "package": "npm run cleanup && ngc && npm run bundling && npm run minify && npm run copy"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/BTMorton/angular2-grid.git"
-  },
-  "keywords": [
-    "angular",
-    "angular2",
-    "grid",
-    "web-components"
-  ],
-  "author": "Ben Morton <ben.morton91@gmail.com> (http://bmorton.co.uk/)",
-  "contributors": [
-    "Martin Sikora (https://github.com/martinsik)",
-    "Daniel Pastor de la Fuenta (https://github.com/aebsubis)",
-    "Frederik Schubert (https://github.com/frederikschubert)",
-    "Matthew de Nobrega (https://github.com/matthewdenobrega)",
-    "Alex (https://github.com/lifecoderua)",
-    "pocmanu (https://github.com/pocmanu)",
-    "Chris Morabito (https://github.com/morabitowoolpert)",
-    "Carlos Esteban Lopez Jaramillo (https://github.com/Luchillo)",
-    "Niklas Berg (https://github.com/nkholski)",
-    "Wouter (https://github.com/wuhkuh)",
-    "Brian Ellis (https://github.com/OnlyAGhost)",
-    "Janne Julkunen (https://github.com/sconix)",
-    "Thomas Schellenberg (https://github.com/ThomasSchellenbergNextCentury)"
-  ],
+  "name": "angular2-grid-demo",
+  "version": "0.0.0",
   "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/BTMorton/angular2-grid/issues"
+  "angular-cli": {},
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "build-prod": "ng build --prod",
+    "build-lib": "ng build --prod angular2-grid && npm run copy-lib",
+    "copy-lib": "copyfiles -f LICENSE README.md projects/angular2-grid/src/*.css dist/angular2-grid",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
   },
-  "homepage": "https://github.com/BTMorton/angular2-grid",
-  "files": [
-    "dist/NgGrid*",
-    "dist/js/",
-    "dist/bundles/",
-    "dist/aot/",
-    "dist/typings/"
-  ],
-  "devDependencies": {
-    "@angular/core": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/compiler-cli": "^5.0.0",
-    "copyfiles": "^1.0.0",
-    "reflect-metadata": "^0.1.8",
-    "rimraf": "^2.5.4",
-    "rollup": "^0.37.0",
-    "rxjs": "^5.5.0",
-    "tslint": "^4.5.1",
-    "typescript": "~2.6.0",
-    "uglify-js": "^2.7.5",
+  "private": true,
+  "dependencies": {
+    "@angular/common": "^6.0.0",
+    "@angular/core": "^6.0.0",
+    "@angular/compiler": "^6.0.0",
+    "@angular/forms": "^6.0.0",
+    "@angular/http": "^6.0.0",
+    "@angular/platform-browser": "^6.0.0",
+    "@angular/platform-browser-dynamic": "^6.0.0",
+    "@angular/router": "^6.0.0",
+    "bootstrap": "^3.3.7",
+    "core-js": "^2.4.1",
+    "rxjs": "^6.1.0",
     "zone.js": "^0.8.4"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~0.6.8",
+    "@angular-devkit/build-ng-packagr": "~0.6.8",
+    "@angular/cli": "^6.0.8",
+    "@angular/compiler-cli": "^6.0.0",
+    "@types/jasmine": "2.5.38",
+    "@types/node": "~6.0.60",
+    "codelyzer": "~2.0.0",
+    "copyfiles": "^2.0.0",
+    "jasmine-core": "~2.5.2",
+    "jasmine-spec-reporter": "~3.2.0",
+    "karma": "~1.4.1",
+    "karma-chrome-launcher": "~2.0.0",
+    "karma-cli": "~1.0.1",
+    "karma-coverage-istanbul-reporter": "^0.2.0",
+    "karma-jasmine": "~1.1.0",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "ng-packagr": "^3.0.0-rc.2",
+    "protractor": "~5.1.0",
+    "ts-node": "~2.0.0",
+    "tsickle": ">=0.25.5",
+    "tslib": "^1.7.1",
+    "tslint": "^4.5.0",
+    "typescript": "~2.7.0"
   }
 }
