+++ projects/angular2-grid/ng-package.json
@@ -1,0 +1,8 @@
+{
+  "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
+  "dest": "../../dist/angular2-grid",
+  "deleteDestPath": false,
+  "lib": {
+    "entryFile": "src/main.ts"
+  }
+}
