+++ projects/angular2-grid/ng-package.prod.json
@@ -1,0 +1,7 @@
+{
+  "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
+  "dest": "../../dist/angular2-grid",
+  "lib": {
+    "entryFile": "src/main.ts"
+  }
+}
