+++ projects/angular2-grid/package.json
@@ -1,0 +1,39 @@
+{
+  "name": "angular2-grid",
+  "version": "3.0.0",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/BTMorton/angular2-grid.git"
+  },
+  "keywords": [
+    "angular",
+    "angular2",
+    "grid",
+    "web-components"
+  ],
+  "author": "Ben Morton <ben.morton91@gmail.com> (http://bmorton.co.uk/)",
+  "contributors": [
+    "Martin Sikora (https://github.com/martinsik)",
+    "Daniel Pastor de la Fuenta (https://github.com/aebsubis)",
+    "Frederik Schubert (https://github.com/frederikschubert)",
+    "Matthew de Nobrega (https://github.com/matthewdenobrega)",
+    "Alex (https://github.com/lifecoderua)",
+    "pocmanu (https://github.com/pocmanu)",
+    "Chris Morabito (https://github.com/morabitowoolpert)",
+    "Carlos Esteban Lopez Jaramillo (https://github.com/Luchillo)",
+    "Niklas Berg (https://github.com/nkholski)",
+    "Wouter (https://github.com/wuhkuh)",
+    "Brian Ellis (https://github.com/OnlyAGhost)",
+    "Janne Julkunen (https://github.com/sconix)",
+    "Thomas Schellenberg (https://github.com/ThomasSchellenbergNextCentury)"
+  ],
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/BTMorton/angular2-grid/issues"
+  },
+  "homepage": "https://github.com/BTMorton/angular2-grid",
+  "peerDependencies": {
+    "@angular/common": "^6.0.0-rc.0 || ^6.0.0",
+    "@angular/core": "^6.0.0-rc.0 || ^6.0.0"
+  }
+}
