+++ projects/angular2-grid/src/main.ts
@@ -1,0 +1,5 @@
+export { NgGrid } from './directives/NgGrid';
+export { NgGridItem } from './directives/NgGridItem';
+export { NgGridPlaceholder } from './components/NgGridPlaceholder';
+export { NgGridConfig, NgGridItemConfig, NgGridItemEvent, NgGridItemSize, NgGridItemPosition, NgGridRawPosition, NgGridItemDimensions } from './interfaces/INgGrid';
+export { NgGridModule } from './modules/NgGrid.module';
