+++ projects/angular2-grid/src/public_api.ts
@@ -1,0 +1,7 @@
+/*
+ * Public API Surface of angular2-grid
+ */
+
+export * from './lib/angular2-grid.service';
+export * from './lib/angular2-grid.component';
+export * from './lib/angular2-grid.module';
