--- src/NgGrid.css
@@ -1,26 +1,0 @@
-.grid {
-	background-color: #efefef;
-	transform-origin: top center;
-	transition: transform 0.5s;
-}
-
-.grid-item {
-	background-color: #ffffff;
-	-webkit-transition: width 0.25s, height 0.25s, transform 0.5s;
-	-moz-transition: width 0.25s, height 0.25s, transform 0.5s;
-	-o-transition: width 0.25s, height 0.25s, transform 0.5s;
-	transition: width 0.25s, height 0.25s, transform 0.5s;
-	overflow: hidden;
-}
-
-.grid-item:active, .grid-item.moving {
-	z-index: 2;
-	-webkit-transition: none;
-	-moz-transition: none;
-	-o-transition: none;
-	transition: none;
-}
-
-.grid-placeholder {
-	background-color: rgba(0, 0, 0, 0.3);
-}
