--- src/NgGrid_FixSmall.css
@@ -1,16 +1,0 @@
-@media (max-width: 767px) {
-	.grid {
-		width: 100% !important;
-		height: auto !important;
-		padding: 10px;
-	}
-	.grid-item {
-		position: static !important;
-		width: 100% !important;
-		margin-bottom: 10px;
-		transform: none !important;
-	}
-	.grid-item:last-child {
-		margin-bottom: 0;
-	}
-}
