+++ src/app/app.component.css
@@ -1,0 +1,16 @@
+.grid-item {
+    border: solid 1px;
+}
+
+.handle {
+    width: 100%;
+    height: 50px;
+    background: blue;
+    color: white;
+    padding: 15px;
+    box-sizing: border-box;
+}
+
+.btn {
+    margin-left: 10px;
+}
