+++ src/main.ts
@@ -1,14 +1,10 @@
-export { NgGrid } from "./directives/NgGrid";
-export { NgGridItem } from "./directives/NgGridItem";
-export { NgGridPlaceholder } from "./components/NgGridPlaceholder";
-export {
-	NgGridConfig,
-	NgGridItemConfig,
-	NgGridItemEvent,
-	NgGridItemSize,
-	NgGridItemPosition,
-	NgGridRawPosition,
-	NgGridItemDimensions,
-	NgConfigFixDirection,
-} from "./interfaces/INgGrid";
-export { NgGridModule } from "./modules/NgGrid.module";
+import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
+import { enableProdMode } from '@angular/core';
+import { environment } from './environments/environment';
+import { AppModule } from './app/app.module';
+
+if (environment.production) {
+  enableProdMode();
+}
+
+platformBrowserDynamic().bootstrapModule(AppModule);
