--- src/modules/NgGrid.module.ts
@@ -1,9 +1,0 @@
-import { NgModule, ModuleWithProviders } from "@angular/core";
-import { NgGrid, NgGridItem, NgGridItemConfig, NgGridItemEvent, NgGridPlaceholder } from "../main";
-
-@NgModule({
-	declarations:     [ NgGrid, NgGridItem, NgGridPlaceholder ],
-	entryComponents:  [ NgGridPlaceholder ],
-	exports:          [ NgGrid, NgGridItem ],
-})
-export class NgGridModule {}
